requirements.txt
flask==3.0.0
kubernetes==28.1.0
docker==6.1.3
schedule==1.2.0
cryptography==42.0.0
prometheus-client==0.19.0
pyOpenSSL==24.0.0
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
gunicorn==21.2.0